giving the integer serializer the xsd:integer URI rather than the xsd:double URI
git-svn-id: http://svn.topazproject.org/svn/head@7688 48088309-b40f-0410-b149-b14953253675

+++ topaz/core/src/main/java/org/topazproject/otm/serializer/SerializerFactory.java
@@ -126,9 +126,9 @@ public class SerializerFactory {
     setSerializer(Boolean.TYPE, new XsdBooleanSerializer(), false);
     setSerializer(Integer.class, new SimpleSerializer<Integer>(Integer.class), false);
     setSerializer(Integer.TYPE, new SimpleSerializer<Integer>(Integer.class), false);
-    setSerializer(Integer.class, Rdf.xsd + "double", new IntegerSerializer<Integer>(Integer.class),
+    setSerializer(Integer.class, Rdf.xsd + "integer", new IntegerSerializer<Integer>(Integer.class),
                   false);
-    setSerializer(Integer.TYPE, Rdf.xsd + "double", new IntegerSerializer<Integer>(Integer.class),
+    setSerializer(Integer.TYPE, Rdf.xsd + "integer", new IntegerSerializer<Integer>(Integer.class),
                   false);
     setSerializer(Long.class, new SimpleSerializer<Long>(Long.class), false);
     setSerializer(Long.TYPE, new SimpleSerializer<Long>(Long.class), false);
